scripts/requirements.txt
# Requirements for curriculum embedding and semantic search scripts

# Option 1: FREE Local Embeddings (RECOMMENDED - No API key needed!)
sentence-transformers>=2.2.0

# Option 2: Google Gemini Embeddings (if you have GEMINI_API_KEY)
google-generativeai>=0.3.0

# Supabase Python client
supabase>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Progress bars for embedding generation
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
